Fix format string warnings in Xcode 4.4

+++ Source/Core/HCBaseDescription.m
@@ -47,7 +47,7 @@
     NSString *description = [value description];
     NSUInteger descriptionLength = [description length];
     if (descriptionLength == 0)
-        [self append:[NSString stringWithFormat:@"<%@: 0x%0x>", NSStringFromClass([value class]), value]];
+        [self append:[NSString stringWithFormat:@"<%@: 0x%0x>", NSStringFromClass([value class]), (unsigned int)value]];
     else if ([description characterAtIndex:0] == '<'
              && [description characterAtIndex:descriptionLength - 1] == '>')
     {
